requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-allauth==0.49.0
django-bootstrap-v5==1.0.11
django-heroku==0.3.1
django-multiselectfield==0.1.12
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
oauth==1.0.1
oauthlib==3.2.0
Pillow==8.0.1
platformdirs==2.5.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.2.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.1
whitenoise==6.0.0
Pillow
python-dotenv
faker
twilio